+++ spec/legion/queue_spec.rb
@@ -3,7 +3,7 @@
 require 'spec_helper'
 require 'legion/transport/queue'
 
-RSpec.describe Legion::Transport::Queue do
+RSpec.describe Legion::Transport::Queue, :rabbitmq do
   it 'is a class' do
     expect(Legion::Transport::Queue).to be_a Class
   end